+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import file_picker
+import flutter_bin
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
+  FlutterBinPlugin.register(with: registry.registrar(forPlugin: "FlutterBinPlugin"))
+}